Broke database code into separate modules
